Changement de nom pour Leadee avec suppression de fichiers inutilisés
