+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:ulearn/screens/loading.dart';
+import 'package:leadee/screens/loading.dart';
+import 'package:leadee/share/app_colors.dart';
 
 void main() => runApp(App());
 
@@ -15,19 +16,4 @@ class App extends StatelessWidget {
       home: LoadingComponent(),
     );
   }
-}
-
-class AppColors {
-  static Color hexToColor(String hexString, {String alphaChannel = 'FF'}) {
-    return Color(int.parse(hexString.replaceFirst('#', '0x$alphaChannel')));
-  }
-
-  AppColors._();
-
-  static Map<int, Color> blue = <int, Color>{
-    50: hexToColor('#006290'),
-    100: hexToColor('#00354F')
-  };
-
-  static Map<int, Color> orange = <int, Color>{50: hexToColor('#D68910')};
 }
