+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: ulearn
+name: leadee
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -46,8 +46,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-    - assets/images/
+  # assets:
+  #   - assets/images/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
