+++ test/widget_test.dart
@@ -8,8 +8,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:ulearn/main.dart';
-
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {});
 }
